Note with same name refactoring across spaces fix (#442)
https://github.com/workduck-io/mex-electron/pull/442

* Note with same name refactoring across spaces fix

* Version bump

+++ src/editor/Actions/useEditorBlockSelection.tsx
@@ -84,9 +84,9 @@ export const useEditorBlockSelection = () => {
       insertNodes(editor, { type: ELEMENT_PARAGRAPH, id: generateTempId(), children: [{ text: '' }] }, { at: [0] })
   }
 
-  const deleteSelectedBlock = (): any => {
+  const deleteSelectedBlock = (deleteBlock?: boolean): any => {
     const editorBlocks = getEditorBlocks()
-    deleteContentBlocks(editorBlocks)
+    if (deleteBlock) deleteContentBlocks(editorBlocks)
     return editorBlocks
   }
